Moved connection and DAQ widgets into separate folder, added ability to stop spectrum acquisition and clear previously obtained data.
